+++ backend/routes/api.js
@@ -1,16 +1,17 @@
 const express = require('express');
-const dbController = require('../controllers/dbController');
+const reviewController = require('../controllers/reviewController');
 const userController = require('../controllers/userController');
 const cookieController = require('../controllers/cookieController');
+const landlordController = require('../controllers/landlordController');
 const auth = require('../middleware/auth');
 const router = express.Router();
 
-router.get('/getall', dbController.getAll);
+router.get('/getall', reviewController.getAll);
 
 router.get(
   '/getlandlord/:id',
-  dbController.getLandLord,
-  dbController.getReviews,
+  landlordController.getLandLord,
+  reviewController.getReviews,
   (req, res) => {
     res.status(200).json({
       landlord: res.locals.landLord,
@@ -19,7 +20,7 @@ router.get(
   }
 );
 
-router.post('/createlandlord', dbController.createLandlord);
+router.post('/createlandlord', landlordController.createLandlord);
 
 router.post(
   '/login',
@@ -34,7 +35,7 @@ router.post('/logout', (req, res) => {
 });
 router.post('/signup', userController.createUsers);
 
-router.post('/postReviews', dbController.postReviews);
-router.post('/postReviews', auth.verifyToken, dbController.postReviews);
+router.post('/postReviews', reviewController.postReviews);
+router.post('/postReviews', auth.verifyToken, reviewController.postReviews);
 
 module.exports = router;
